+++ engine/lib/membrane_rtc_engine/engine.ex
@@ -378,6 +378,7 @@ defmodule Membrane.RTC.Engine do
         ) :: :ok | {:error, :timeout | :invalid_track_id}
   def subscribe(rtc_engine, endpoint_id, track_id, opts \\ []) do
     ref = make_ref()
+
     send(rtc_engine, {:subscribe, {self(), ref}, endpoint_id, track_id, opts})
 
     receive do
@@ -481,7 +482,7 @@ defmodule Membrane.RTC.Engine do
       :ok ->
         {spec, state} = fulfill_or_postpone_subscription(subscription, ctx, state)
         send(endpoint_pid, {ref, :ok})
-        Membrane.Logger.info("Subscription fullfiled by #{endpoint_id} on track: #{track_id}")
+        Membrane.Logger.debug("Subscription fullfiled by #{endpoint_id} on track: #{track_id}")
         {[spec: {spec, log_metadata: [rtc: state.id]}], state}
 
       {:error, _reason} = error ->