+++ engine/lib/membrane_rtc_engine/exception.ex
@@ -38,24 +38,6 @@ defmodule Membrane.RTC.Engine.Exception do
     end
   end
 
-  defmodule RequestTrackVariantError do
-    defexception [:message]
-
-    @impl true
-    def exception(opts) do
-      requester = Keyword.fetch!(opts, :requester)
-      requested_variant = Keyword.fetch!(opts, :requested_variant)
-      track = Keyword.fetch!(opts, :track)
-
-      msg = """
-      Endpoint: #{inspect(requester)} requested non-existing track variant: \
-      #{inspect(requested_variant)} for track: #{inspect(track)}.
-      """
-
-      %__MODULE__{message: msg}
-    end
-  end
-
   defmodule TrackVariantStateError do
     defexception [:message]